requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
PyYAML==6.0.2
google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
